# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/repair-requests": {
            "get": {
                "tags": [
                    "Repair Requests"
                ],
                "operationId": "6927b3f4307110bc53bd91859ee37381",
                "responses": {
                    "200": {
                        "description": "Request list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Repair Requests"
                ],
                "operationId": "bcf5ace619aa690c88f314b5b7546bea",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/RepairRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Request created"
                    }
                }
            }
        },
        "/api/v1/repair-requests/{id}": {
            "put": {
                "tags": [
                    "Repair Requests"
                ],
                "operationId": "5c24b4b97cba46b077a29f9c6f71294e",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/RepairRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/repair-requests/{id}/assign-technician": {
            "put": {
                "tags": [
                    "Repair Requests"
                ],
                "operationId": "47f55f726f48c5170609984f05f67e71",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "technician_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Technician assigned"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/repair-requests/{id}/status": {
            "patch": {
                "tags": [
                    "Repair Requests"
                ],
                "operationId": "3a241457b165b69bb14b2bc3b4248cc7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RepairRequest": {
                "title": "RepairRequest",
                "required": [
                    "id",
                    "description",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "description": {
                        "type": "string",
                        "example": "Device not working"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}